refact : add model and move class
