+++ src/test/java/racingcar/model/CarTest.java
@@ -1,10 +1,12 @@
-package racingcar;
+package racingcar.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.model.Car;
+import racingcar.model.CarMovementStatus;
 
 public class CarTest {
     Car car;
